+++ frontends/web/src/routes/account/actionButtons.tsx
@@ -7,9 +7,10 @@ import { ArrowFloorDownWhite, ArrowFloorUpWhite, Coins, WalletConnectLight } fro
 import { useMediaQuery } from '@/hooks/mediaquery';
 import { AccountCode, TAccount, CoinCode } from '@/api/account';
 import { isEthereumBased } from './utils';
-import { connectKeystore } from '@/api/keystores';
 import { AccountActionButtonLink } from './components/account-action-button-link';
 import { AccountActionButtons } from './components/account-action-buttons';
+import { FirmwareUpgradeRequiredDialog } from '@/components/dialog/firmware-upgrade-required-dialog';
+import { useFeatureConnect } from '@/hooks/keystore';
 import style from './account.module.css';
 
 type TProps = {
@@ -24,18 +25,25 @@ type TProps = {
 export const ActionButtons = ({ canSend, code, coinCode, exchangeSupported, account, accountDataLoaded }: TProps) => {
   const { t } = useTranslation();
   const navigate = useNavigate();
+  const {
+    connect,
+    dismissFirmwareUpgrade,
+    firmwareUpgradeRequired,
+  } = useFeatureConnect();
   const walletConnectEnabled = isEthereumBased(account.coinCode) && !account.isToken;
   const isLargeTablet = useMediaQuery('(max-width: 830px)');
   const isMobile = useMediaQuery('(max-width: 768px)');
 
-  // When clicking 'Send', for Ethereum based accounts we first prompt to connect the keystore
-  // before proceeding. The reason is that in ETH, we need to know what keystore (which BitBox02
-  // version) is connected to decide which ETH transaction proposals to construct (legacy vs EIP1559).
+  // When clicking 'Send', first prompt to connect the keystore and check that it supports signing.
+  // For Ethereum based accounts, we also need to know which keystore (which BitBox02 version) is
+  // connected to decide which ETH transaction proposals to construct (legacy vs EIP1559).
   const sendLink = `/account/${code}/send`;
-  const maybeRouteSend = async (e: MouseEvent<HTMLAnchorElement>) => {
+  const routeSend = async (e: MouseEvent<HTMLAnchorElement>) => {
     e.preventDefault();
-    const connectResult = await connectKeystore(account.keystore.rootFingerprint);
-    if (connectResult.success) {
+    const requiredFeature = isEthereumBased(coinCode)
+      ? 'ethTransactionSigning'
+      : 'btcTransactionSigning';
+    if (await connect(account.keystore.rootFingerprint, requiredFeature)) {
       // Proceed to the send screen if the keystore was connected.
       navigate(sendLink);
     }
@@ -48,7 +56,7 @@ export const ActionButtons = ({ canSend, code, coinCode, exchangeSupported, acco
       <AccountActionButtonLink
         disabled={!canClickSend}
         to={sendLink}
-        onClick={isEthereumBased(coinCode) ? maybeRouteSend : undefined}
+        onClick={routeSend}
       >
         <ArrowFloorUpWhite width={16} height={16} />
         <span>{t('generic.send')}</span>
@@ -83,6 +91,12 @@ export const ActionButtons = ({ canSend, code, coinCode, exchangeSupported, acco
             <span>Wallet Connect</span>
           )}
         </AccountActionButtonLink>
+      )}
+      {firmwareUpgradeRequired && (
+        <FirmwareUpgradeRequiredDialog
+          open
+          onClose={dismissFirmwareUpgrade}
+        />
       )}
     </AccountActionButtons>
   );