+++ frontends/web/src/routes/account/sign-message/sign-message.tsx
@@ -14,6 +14,7 @@ import {
 import { useSignMessageController } from './use-sign-message-controller';
 import { isBitcoinBased } from '../utils';
 import { AddressesContent } from '../addresses/addresses';
+import { FirmwareUpgradeRequiredDialog } from '@/components/dialog/firmware-upgrade-required-dialog';
 import styles from './sign-message.module.css';
 
 type TProps = {
@@ -42,6 +43,12 @@ export const SignMessage = ({
 
   return (
     <Main>
+      {controller.firmwareUpgradeRequired && (
+        <FirmwareUpgradeRequiredDialog
+          open
+          onClose={controller.dismissFirmwareUpgrade}
+        />
+      )}
       <Header
         hideSidebarToggler
         title={